pyproject.toml
[tool.poetry]
authors = ["Patrick Forringer <patrick@forringer.com>"]
description = "Inventory management system"
license = "MIT"
name = "invo"
version = "0.1.0"

[tool.poetry.dependencies]
django = "^3.1.3"
django-extensions = "^3.0.9"
django-mptt = "^0.11.0"
django-mptt-admin = "^2.0.0"
django-polymorphic = "^3.0.0"
django-polymorphic-tree = "^1.5.1"
django-safedelete = "^0.5.6"
psycopg2-binary = "^2.8.6"
pyhumps = "^1.6.1"
python = "^3.9"
ariadne-extended = {path = "../python/ariadne-extended"}
django-filter = "^2.4.0"
django-waffle = "^2.0.0"
dateutils = "^0.6.11"
django-fsm = "^2.7.0"
arrow = "^0.17.0"
django-cors-headers = "^3.5.0"
glom = "^20.11.0"
django-configurations = {version = "^2.2", extras = ["cache", "database", "email", "search"]}
django-haystack = "^3.0.0"
elasticsearch = "^5.0.0"
segno = "^1.3.1"
django-measurement = {git = "https://github.com/coddingtonbear/django-measurement.git", rev = "v4"}
marshmallow = "^3.9.1"
django-model-utils = "^4.1.1"
rich = "^9.5.1"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
ipdb = "^0.13.4"
model_bakery = "^1.2.0"
pytest-django = "^3.10.0"
pytest-mock = "^3.3.1"
pytest-watch = "^4.2.0"
mock = "^4.0.2"
pylint = "^2.6.0"
pytest = "^6.1.2"
flake8 = "^3.8.4"
flakehell = "^0.7.1"
rope = "^0.18.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "invo.config"
DJANGO_CONFIGURATION = "Test"
addopts = "--no-migrations --disable-pytest-warnings -vvv"
python_files = "tests.py test_*.py *_tests.py"

[tool.black]
exclude = '/(\.eggs|\.git|\.mypy_cache|\.tox|\.venv|build|dist|migrations)/'
include = '\.pyi?$'
line-length = 99
target-version = ['py38']

[tool.flake8]
format = "colored"
max_line_length = 99

# [tool.flakehell.plugins]
# flake8-isort = ["+*"]

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
